Commented out demo line
People compiling this were only getting it to say "135 bpm" with no change.  This was cause I wanted to take some screen shots and forced the text to only say that.  I forgot to remove the line after taking the screenshots.

Commenting out this line now for anyone who wants to compile this on their own.